chore(edit-content): add docs #29875
